feat(tasks): Add protos for the ConsumerService
Add proto definitions for the ConsumerService as well as protos for the messages used
to interact with that service.

Add a common error structure for returning errors.

The service currently has support for GetTask and SetTaskStatus, which is what the workers
will use to fetch new tasks and update the task statuses after a task is finished.